keyboard - rename and cleanup some editor related contextkeys
